C11 ex05 rendered

+++ pisine/C11/ex05/srcs/ft_op.c
@@ -6,31 +6,31 @@
 /*   By: hseong <hseong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/25 20:31:48 by hseong            #+#    #+#             */
-/*   Updated: 2021/10/25 20:31:50 by hseong           ###   ########.fr       */
+/*   Updated: 2021/10/25 21:14:56 by hseong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long long	ft_long_add(long a, long b)
+long long	ft_long_add(long long a, long long b)
 {
 	return (a + b);
 }
 
-long long	ft_long_sub(long a, long b)
+long long	ft_long_sub(long long a, long long b)
 {
 	return (a - b);
 }
 
-long long	ft_long_mul(long a, long b)
+long long	ft_long_mul(long long a, long long b)
 {
 	return (a * b);
 }
 
-long long	ft_long_div(long a, long b)
+long long	ft_long_div(long long a, long long b)
 {
 	return (a / b);
 }
 
-long long	ft_long_mod(long a, long b)
+long long	ft_long_mod(long long a, long long b)
 {
 	return (a % b);
 }
